Make <FeaturedProduct> component presentational
